+++ experiments/classification_active_learning.py
@@ -19,6 +19,7 @@ from alpaca.dataloader.builder import build_dataset
 from alpaca.uncertainty_estimator.masks import DEFAULT_MASKS
 from alpaca.active_learning.simple_update import update_set
 from utils.fastai import ImageArrayDS
+from utils.visual_datasets import prepare_mnist, prepare_cifar, prepare_svhn
 
 if torch.cuda.is_available():
     torch.cuda.set_device(1)
@@ -113,50 +114,6 @@ def build_model(model_type):
     return model
 
 
-def prepare_cifar(config):
-    dataset = build_dataset('cifar_10', val_size=config['val_size'])
-    x_set, y_set = dataset.dataset('train')
-    x_val, y_val = dataset.dataset('val')
-
-    shape = (-1, 3, 32, 32)
-    x_set = ((x_set - 128) / 128).reshape(shape)
-    x_val = ((x_val - 128) / 128).reshape(shape)
-
-    train_tfms = [*rand_pad(4, 32), flip_lr(p=0.5)]  # Transformation to augment images
-
-    return x_set, y_set, x_val, y_val, train_tfms
-
-
-def prepare_mnist(config):
-    dataset = build_dataset('mnist', val_size=config['val_size'])
-    x_set, y_set = dataset.dataset('train')
-    x_val, y_val = dataset.dataset('val')
-
-    shape = (-1, 1, 28, 28)
-    x_set = ((x_set - 128) / 128).reshape(shape)
-    x_val = ((x_val - 128) / 128).reshape(shape)
-
-    train_tfms = []
-
-    return x_set, y_set, x_val, y_val, train_tfms
-
-
-def prepare_svhn(config):
-    dataset = build_dataset('svhn', val_size=config['val_size'])
-    x_set, y_set = dataset.dataset('train')
-    x_val, y_val = dataset.dataset('val')
-    y_set[y_set == 10] = 0
-    y_val[y_val == 10] = 0
-
-    shape = (-1, 32, 32, 3)
-    x_set = ((x_set - 128) / 128).reshape(shape)
-    x_val = ((x_val - 128) / 128).reshape(shape)
-    x_set = np.rollaxis(x_set, 3, 1)
-    x_val = np.rollaxis(x_val, 3, 1)
-
-    train_tfms = [*rand_pad(4, 32), flip_lr(p=0.5)]  # Transformation to augment images
-
-    return x_set, y_set, x_val, y_val, train_tfms
